--- ELDA/_tb_da_clustering_master.json
@@ -1,1 +1,0 @@
-[["2017-01-17T11:42:33Z","2016-12-07","2016-12-07",15,"","0002.0000003F:0002.0000002A:0001.00000030:0002.00000020:0002.0000002D:0002.00000022:0001.00000009:0002.0000002C:0002.00000029:0002.00000038:0002.00000036:0002.00000037:0002.0000003E:0002.00000034:0001.0000000A:0002.00000039:0001.0000000B:0002.00000033:0001.00000006:0002.00000035:0001.0000001D:0001.00000004:0001.00000005:0001.0000001E:0002.00000027","0001.0000001A:0001.00000019:0001.00000018:0001.00000043","0001.0000001E:0002.0000001F:0002.00000024:0001.00000002:0002.00000023:0001.00000001:0001.00000032:0001.0000002F:0002.0000002D:0001.0000000B:0002.00000026:0002.00000029:0002.00000020:0002.0000002C:0002.00000033:0002.00000036:0002.0000002E:0001.00000031:0002.00000038:0001.00000009:0002.00000037:0002.0000003E:0002.00000022:0002.00000035:0001.0000000A:0001.00000005:0002.00000039:0002.00000034:0002.00000028:0001.00000030:0002.0000003F:0001.00000044:0002.0000002A:0001.0000001C:0001.0000001D:0001.00000006:0001.00000004:0002.00000027:0001.00000007:0001.0000003A"]]
