+++ Makefile
@@ -2,4 +2,4 @@ init:
 	pip install -r requirements.txt
 
 run:
-	python matplotlib2.py
+	python elda_main.py
