+++ requirements.txt
@@ -1,8 +1,3 @@
-jpype
-jaydebeapi
 pandas
+requests
 numpy
-matplotlib
-sklearn
-pandasql
-flask-restful
