--- tb_da_clustering_master.txt
@@ -1,2 +1,0 @@
-da_date	s_date	e_date	time_interval	c0_voltage_node	c1_voltage_node	c2_voltage_node	c3_voltage_node	c0_ampere_node	c1_ampere_node	c2_ampere_node	c3_ampere_node	c0_active_power_node	c1_active_power_node	c2_active_power_node	c3_active_power_node	c0_power_factor_node	c1_power_factor_node	c2_power_factor_node	c3_power_factor_node
-2017-01-11 13:33:32.812572	2016-12-01	2016-12-02	15	"0002.00000029", "0002.00000028", "0002.0000003E", "0002.0000002C", "0002.00000020", "0002.0000002D", "0002.0000002A", "0002.0000003F", "0002.0000002E", "0002.00000022", "0002.00000023", "0001.0000002F", "0002.00000025", "0002.0000001F", "0001.00000031", "0002.00000027", "0001.00000030", "0002.00000024", "0002.00000026", "0001.00000032"	"0001.00000013", "0001.00000012", "0001.00000011", "0001.00000040", "0001.0000000E", "0001.0000000D", "0001.0000000C"	"0001.0000001C", "0001.0000001D", "0001.00000043", "0001.00000006", "0001.00000005", "0001.00000004", "0001.00000007", "0001.0000000A", "0001.00000009", "0001.0000003A", "0002.00000034", "0002.00000036", "0002.00000033", "0001.00000001", "0001.0000000B", "0001.00000002", "0002.00000037", "0002.00000038", "0002.00000039", "0001.00000017", "0001.00000018", "0001.00000019", "0001.0000001A", "0001.00000016"	"0001.00000015"	"0001.00000040", "0002.00000026", "0001.00000013", "0002.0000002A", "0001.0000002F", "0002.00000027", "0002.0000002D", "0002.00000038", "0001.00000030", "0002.0000002C", "0002.00000036", "0002.00000037", "0002.00000034", "0002.00000033", "0001.0000001D", "0001.00000043", "0001.00000005"	"0002.00000023", "0001.00000002", "0002.0000001F", "0001.00000001", "0001.00000018", "0001.00000019", "0001.00000032", "0001.00000007", "0001.0000000E", "0001.00000011"	"0001.00000015"	"0002.00000024", "0002.00000028", "0002.00000025", "0001.0000000C", "0001.0000000D", "0001.00000031", "0001.0000001C", "0001.00000012", "0001.00000014", "0002.0000002E", "0002.0000003F", "0001.00000009", "0002.00000020", "0002.00000029", "0002.00000022", "0002.0000003E", "0002.00000039", "0001.00000017", "0001.00000016", "0001.0000000A", "0001.0000000B", "0001.00000006", "0001.00000004", "0001.0000001E"	"0001.00000015"	"0002.00000023", "0001.00000019", "0002.0000001F", "0001.00000018", "0001.00000044", "0002.0000002A", "0001.00000011", "0002.00000034", "0002.00000026", "0001.0000000E", "0002.0000002D", "0002.00000037", "0001.0000001D", "0002.00000027", "0001.00000043", "0001.0000002F", "0001.00000005", "0002.00000033", "0001.00000030"	"0001.00000032", "0001.0000000C", "0001.00000014", "0001.0000000A", "0001.00000004", "0002.0000003E", "0002.0000002E", "0001.0000001E", "0002.0000003F", "0002.00000022", "0001.00000017"	"0001.0000001A", "0002.00000024", "0002.00000028", "0002.00000025", "0001.00000007", "0001.0000001C", "0001.00000009", "0002.00000036", "0001.00000031", "0001.00000040", "0002.0000002C", "0002.00000029", "0002.00000038", "0002.00000039", "0001.0000000D", "0001.0000000B", "0001.00000013", "0001.00000006", "0001.00000012", "0002.00000020", "0001.00000016"	"0001.0000003A", "0001.00000040", "0001.00000007", "0002.00000036", "0001.00000013", "0002.00000038", "0002.0000002C", "0002.00000033", "0002.00000026"	"0002.00000035", "0001.00000044", "0001.00000043"	"0001.00000014", "0001.0000000C", "0002.00000024", "0002.00000028", "0002.00000025", "0001.0000000D", "0002.0000003E", "0001.00000017", "0002.00000022", "0001.0000001C", "0002.00000029", "0001.00000031", "0002.0000002E", "0002.0000003F", "0001.00000004", "0002.00000039", "0002.00000020", "0001.0000000B", "0001.00000016", "0001.00000006", "0001.00000009", "0001.00000012"	"0002.00000023", "0001.00000001", "0002.0000001F", "0001.00000032", "0001.00000018", "0001.00000019", "0001.00000011", "0001.00000002", "0002.00000027", "0002.0000002D", "0001.0000001D", "0001.0000002F", "0002.0000002A", "0002.00000037", "0001.0000000A", "0001.00000005", "0001.0000000E", "0001.00000030"
